Final Commit

+++ src/main/java/com/asish/springbootqrcode/model/student/Student.java
@@ -19,12 +19,12 @@ public class Student {
     private String firstName;
     private  String lastName;
     private String email;
-    private Integer mobile;
+    private Long mobile;
 
     public Student(String firstName, String lastName, String email, Integer mobile) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
-        this.mobile = mobile;
+        this.mobile = Long.valueOf(mobile);
     }
 }
